requirements.txt
# Web interface
streamlit>=1.28.0

# Document processing
python-docx>=0.8.11
openpyxl>=3.1.2
xlrd>=2.0.1

# Text matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
